+++ bank-feeds/src/sdk/models/operations/updatesourceaccount.ts
@@ -11,19 +11,28 @@ export class UpdateSourceAccountRequest extends SpeakeasyBase {
     sourceAccount?: shared.SourceAccount;
 
     /**
-     * Unique identifier for an account
+     * Unique identifier for an account.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=accountId" })
     accountId: string;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 }
 
 export class UpdateSourceAccountResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -39,9 +48,15 @@ export class UpdateSourceAccountResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     sourceAccount?: shared.SourceAccount;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }